+++ lib/lifecycle.dart
@@ -1,3 +1,4 @@
+export 'src/lifecycle/builder.dart';
+export 'src/lifecycle/live_data.dart';
 export 'src/lifecycle/view_model.dart';
-export 'src/lifecycle/view_model_builder.dart';
-export 'src/lifecycle/view_model_scope.dart';
+export 'src/lifecycle/view_model_state.dart';