+++ lib/src/lifecycle/view_model.dart
@@ -1,92 +1,21 @@
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/foundation.dart';
-
-import 'package:flutter_boot/src/http/an_http.dart';
-import 'package:flutter_boot/src/http/an_param.dart';
+import 'package:flutter_boot/http.dart';
 import 'package:flutter_boot/widget.dart';
-import 'view_model_scope.dart';
+import 'package:flutter_boot/lifecycle.dart';
 
-/// value 上一次的值
-typedef ViewModelStateCallBack<T> = void Function(T value);
 
-/// 状态值
-class ViewModelState<T> extends ChangeNotifier implements ValueListenable<T> {
-  ViewModelState._(this._value, {bool notify = false}) {
-    if (notify) {
-      notifyListeners();
-    }
-  }
+class ActionViewModel extends ViewModel {
+  Function(
+    String message, {
+    int duration,
+    ToastAlignment alignment,
+    Widget? widget,
+  })? toastCall;
 
-  bool hostDispose = false;
+  Function({Widget? widget})? showLoadingCall;
 
-  @override
-  T get value => _value;
-
-  T _value;
-
-  set value(T newValue) {
-    if (_value == newValue) {
-      return;
-    }
-    _value = newValue;
-    // 为增加更新的可控制性，要主动去调用
-    // notifyListeners();
-  }
-
-  /// 通知更新
-  void _setState<V>(ViewModelStateCallBack<V> fn) {
-    if (hostDispose) {
-      return;
-    }
-    fn(value as V);
-    this.notifyListeners();
-  }
-
-  /// 不建议直接调用。请使用 ViewModel setState()方法。
-  @override
-  void notifyListeners() {
-    if (hostDispose) {
-      return;
-    }
-    super.notifyListeners();
-  }
-
-  @override
-  String toString() => '$this->value:($value)';
-}
-
-/// Live ViewModel
-class LiveViewModel<S extends ViewModelScope> {
-  /// 所持有的视图层
-  final S _scope;
-
-  LiveViewModel(this._scope);
-
-  /// 创建 ViewModelState
-  /// notify 创建时是否更新状态
-  ViewModelState<T> createState<T>(T value, {bool notify = false}) {
-    var valueNotifier = ViewModelState._(value, notify: notify);
-    _scope.putViewModelStateNotifier(valueNotifier);
-    return valueNotifier;
-  }
-
-  void setScopeState(VoidCallback fn) {
-    _scope.setState(fn);
-  }
-
-  void setState<T>(ViewModelState valueNotifier, ViewModelStateCallBack<T> fn) {
-    valueNotifier._setState<T>(fn);
-  }
-
-  void sendNotify(String message, {int? what, Object? data}) {
-    _scope.onNotify(message, what: what, data: data);
-  }
-}
-
-/// Action ViewModel
-class ActionViewModel<S extends ActionViewModelScope> extends LiveViewModel<S> {
-  ActionViewModel(super.scope);
+  VoidCallback? dismissLoadingCall;
 
   void toast(
     String message, {
@@ -94,54 +23,109 @@ class ActionViewModel<S extends ActionViewModelScope> extends LiveViewModel<S> {
     ToastAlignment alignment = ToastAlignment.bottom,
     Widget? widget,
   }) {
-    _scope.toast(message,
-        duration: duration, alignment: alignment, widget: widget);
+    if (toastCall != null) {
+      toastCall!(message,
+          duration: duration, alignment: alignment, widget: widget);
+    }
   }
 
   void showLoading({Widget? widget}) {
-    _scope.showLoading(widget: widget);
+    if(showLoadingCall!=null) {
+      showLoadingCall!(widget: widget);
+    }
+
   }
 
   void dismissLoading() {
-    _scope.dismissLoading();
+    if(dismissLoadingCall!=null){
+      dismissLoadingCall!();
+    }
+
   }
 }
 
-/// Http ViewModel
-class HttpViewModel<S extends HttpViewModelScope> extends ActionViewModel<S> {
-  HttpViewModel(S scope) : super(scope);
+class HttpViewModel extends ActionViewModel {
+  final List<CancelToken> _httpRequestTokens = [];
 
+  List<CancelToken> get httpRequestTokens => _httpRequestTokens;
+
+  void putHttpRequestToken(CancelToken cancelToken) {
+    _httpRequestTokens.add(cancelToken);
+  }
+
+  /// dio 原始返回类型 请求
   Future<Response<T>> anHttp<T>(Param param,
       {HttpMethodType method = HttpMethodType.post,
       ProgressCallback? onSendProgress,
       ProgressCallback? onReceiveProgress}) {
-    return _scope.anHttp<T>(param,
+    var cancelToken = HttpRequestToken();
+    putHttpRequestToken(cancelToken);
+
+    return AnHttp.anHttp<T>(param,
         method: method,
+        cancelToken: cancelToken,
         onSendProgress: onSendProgress,
         onReceiveProgress: onReceiveProgress);
+
+    // return AnHttp.instance
+    //     .request<T>(param, method: method, cancelToken: cancelToken);
   }
 
+  /// http请求， 用于返回值是Object的请求。
+  /// convert 数据类型转换器。
   Future<T> anHttpJson<T>(Param param,
       {required JsonObjectConvertor<T> convertor,
       HttpMethodType method = HttpMethodType.post,
       ProgressCallback? onSendProgress,
       ProgressCallback? onReceiveProgress}) async {
-    return _scope.anHttpJson<T>(param,
-        method: method,
+    var cancelToken = HttpRequestToken();
+    putHttpRequestToken(cancelToken);
+
+    // var response = await AnHttp.anHttp<Map<String, dynamic>>(param,
+    //     method: method,
+    //     cancelToken: cancelToken,
+    //     onSendProgress: onSendProgress,
+    //     onReceiveProgress: onReceiveProgress);
+
+    // var response = await AnHttp.instance.request<Map<String, dynamic>>(param,
+    //     method: method, cancelToken: cancelToken);
+
+    // return convertor(response.data!);
+    var response = await AnHttp.anHttpJson<T>(param,
         convertor: convertor,
+        cancelToken: cancelToken,
         onSendProgress: onSendProgress,
         onReceiveProgress: onReceiveProgress);
+    return response;
   }
 
-  Future<List<T>> anHttpArray<T>(Param param,
-      {required JsonArrayConvertor<T> convertor,
-      HttpMethodType method = HttpMethodType.post,
-      ProgressCallback? onSendProgress,
-      ProgressCallback? onReceiveProgress}) async {
-    return _scope.anHttpArray<T>(param,
-        method: method,
+  /// http请求， 用于返回值是数组的请求。
+  /// convert 数据类型转换器。
+  Future<List<T>> anHttpArray<T>(
+    Param param, {
+    required JsonArrayConvertor<T> convertor,
+    HttpMethodType method = HttpMethodType.post,
+    ProgressCallback? onSendProgress,
+    ProgressCallback? onReceiveProgress,
+  }) async {
+    var cancelToken = HttpRequestToken();
+    putHttpRequestToken(cancelToken);
+    // var response = await AnHttp.instance.request<List<dynamic>>(param,
+    //     method: method, cancelToken: cancelToken);
+
+    // var response = await AnHttp.anHttp<List<dynamic>>(param,
+    //     method: method,
+    //     cancelToken: cancelToken,
+    //     onSendProgress: onSendProgress,
+    //     onReceiveProgress: onReceiveProgress);
+    // return convertor(response.data!);
+
+    var response = await AnHttp.anHttpArray<T>(param,
         convertor: convertor,
+        cancelToken: cancelToken,
         onSendProgress: onSendProgress,
         onReceiveProgress: onReceiveProgress);
+    return response;
   }
 }
+